opsclaw::tools

Module discovery

Discovery scan engine for mapping what is running on a target host.

Runs a set of standard commands via a CommandRunner and parses their output into a TargetSnapshot that captures OS info, containers, services, listening ports, disk, memory, and load.

Structs

CommandOutput

Output of a single command execution.

ContainerInfo

DiskInfo

K8sDeployment

K8sNode

K8sPod

K8sService

KubernetesInfo

LoadInfo

MemoryInfo

OsInfo

PortInfo

ServiceInfo

TargetSnapshot

Traits

CommandRunner

Abstraction over how commands are executed (SSH, local shell, mock, …).

Functions

discover_kubernetes

parse_df

parse_docker_ps_json

parse_free

parse_k8s_deployments_json

parse_k8s_nodes_json

parse_k8s_pods_json

parse_k8s_services_json

parse_os_release

parse_ps_aux

parse_ss

parse_systemctl

parse_uname

parse_uptime

run_discovery_scan

snapshot_to_markdown

split_addr_port 🔒